CONSIGNA 2 COMPLETADA

+++ prueba-service/src/main/java/com/utn/frc/backend/pruebaservice/dtos/EmpleadoDTO.java
@@ -12,6 +12,6 @@ import java.util.List;
 public class EmpleadoDTO {
     private String nombre;
     private String apellido;
-    private Integer telefono;
+    private Long telefono;
     private List<Integer> pruebasId;
 }
